+++ src/table/expandable-rows/styles.scss
@@ -6,7 +6,6 @@
 @use '../../internal/styles/index' as styles;
 @use '../../internal/styles/tokens' as awsui;
 @use '../../internal/hooks/focus-visible' as focus-visible;
-@use '../../internal/styles/tokens' as tokens;
 
 @use './motion';
